Moved http to imi module, unit tests 100%
