+++ imi/src/main/java/org/motechproject/nms/imi/service/SmsNotificationService.java
@@ -1,0 +1,13 @@
+package org.motechproject.nms.imi.service;
+
+/**
+ * Initiate sms notification with IMI
+ */
+public interface SmsNotificationService {
+
+    /**
+     * Used to initiate sms workflow with IMI
+     * @param callingNumber phone number to send sms to
+     */
+    boolean sendSms(Long callingNumber);
+}
