+++ imi/src/main/resources/META-INF/motech/applicationContext.xml
@@ -26,6 +26,11 @@
                 <value>classpath:imi.properties</value>
             </list>
         </property>
+        <property name="rawConfigFiles">
+            <list>
+                <value>classpath:smsTemplate.json</value>
+            </list>
+        </property>
     </bean>
 
 </beans>
