+++ imi/src/main/resources/imi.properties
@@ -60,3 +60,15 @@ imi.scp.to_command=/bin/cp {src} {dst}
 
 #timeout (in ms) for the scp command
 imi.scp.timeout=60000
+
+## SMS settings ##
+# TODO: replace nmssenderid with a real production sender id
+imi.sms.sender.id = nmsfoobar
+
+# TODO: replace localhost with a real service url
+imi.sms.notification.url = http://localhost/smsmessaging/v1/outbound/nmssenderid/requests
+
+# TODO: figure out where to get the sms content from (already emailed bbc)
+imi.sms.course.completion.message = This is a foobar message that needs to be replaced
+
+imi.sms.status.callback.url = http://localhost/api/mobileacademy/smsdeliverystatus
