+++ mobile-academy/src/main/resources/META-INF/motech/applicationContext.xml
@@ -12,14 +12,8 @@
     <mvc:annotation-driven />
 
     <bean id="maSettings" class="org.motechproject.server.config.SettingsFacade" autowire="byType">
-        <property name="configFiles">
-            <list>
-                <value>classpath:ma-imi.properties</value>
-            </list>
-        </property>
         <property name="rawConfigFiles">
             <list>
-                <value>classpath:smsTemplate.json</value>
                 <value>classpath:nmsCourse.json</value>
             </list>
         </property>
